backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.7.2
google-cloud-firestore==2.16.0
gunicorn==22.0.0
